fix: string timestamps instead of int
Loki push API takes timestamps and log lines. Timestamps were always supposed to be
strings, but the API worked with integers in the JSON payload. This is not the case
anymore and this commit fixes that.

As a bonus, timestamps are now the timestamps of the actual tests instead of time.now().
This bumps the required pytest version to >=7.3.

Signed-off-by: Q. Anglade <quentin.anglade@easymile.com>